Move W/xbar utilities into mpisppy.utils.w_utils subpackage
Groups the W/xbar serialization and tracking modules (wxbarutils,
wxbarreader, wxbarwriter, wxbartester, wtracker) into a new
mpisppy/utils/w_utils/ subpackage, and updates all internal imports,
tests, and docs accordingly. Also adds an explicit
mpisppy.utils.w_utils.wxbarutils import in cylinders/lagranger_bounder.py
that was previously implicit.

Drive-by: update README tagline from "Pyomo models" to "Pyomo and
other models" to reflect the agnostic (AMPL/GAMS/gurobipy) support.
