+++ mpisppy/tests/examples/w_test_data/farmer_set.py
@@ -19,8 +19,8 @@ import mpisppy.utils.cfg_vanilla as vanilla
 
 from mpisppy.extensions.norm_rho_updater import NormRhoUpdater
 from mpisppy.convergers.norm_rho_converger import NormRhoConverger
-from mpisppy.utils.wxbarwriter import WXBarWriter
-from mpisppy.utils.wxbarreader import WXBarReader
+from mpisppy.utils.w_utils.wxbarwriter import WXBarWriter
+from mpisppy.utils.w_utils.wxbarreader import WXBarReader
 from mpisppy.extensions.extension import MultiExtension
 
 write_solution = True
